backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for FlickFindr"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "psycopg2-binary",
    "sqlalchemy",
    "redis",
    "celery",
    "pydantic",
    "pydantic-settings",
    "loguru",
    "flower",
    "pandas",
    "python-dotenv",
    "sentence-transformers>=5.2.0",
    "pgvector>=0.4.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
